python/requirements.txt
aiocache==0.11.1
aiodns==3.0.0
aiofiles==0.8.0
aiohttp==3.8.1
aiohttp-client-cache==0.6.1
aiosignal==1.2.0
aiosqlite==0.17.0
anyio==3.5.0
appdirs==1.4.4
asgiref==3.5.1
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
Brotli==1.0.9
cattrs==1.10.0
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
Deprecated==1.2.13
fastapi==0.75.2
frozenlist==1.3.0
h11==0.13.0
idna==3.3
iidda-api==0.0.1
itsdangerous==2.1.2
multidict==6.0.2
nest-asyncio==1.5.5
numpy==1.22.3
pandas==1.4.2
pipdeptree==2.2.1
pycares==4.1.2
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
PyGithub==1.55
PyJWT==2.3.0
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-forge==18.6.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-cache==0.9.4
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
toml==0.10.2
typing_extensions==4.2.0
url-normalize==1.4.3
urllib3==1.26.9
uvicorn==0.17.6
websockets==10.3
wrapt==1.14.1
yarl==1.7.2